requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
cohere==4.18.0
datasets==2.16.1
docker==7.0.0
errant==2.3.3
hydra-core==1.3.2
ipykernel==6.25.0
ipython==8.14.0
jsonlines==3.1.0
mosestokenizer==1.2.1
openai==0.27.8
pandas==2.2.0
protobuf==4.24.2
python-dotenv==1.0.0
pynvml==11.5.0
seaborn=0.13.2
sentencepiece==0.1.99
scipy==1.11.2
tenacity==8.2.3
tiktoken==0.4.0
transformers==4.32.1